started extracting from main file
