+++ code/whack-a-pirate.py
@@ -5,6 +5,7 @@ import sys
 import numpy as np
 #import psychopy
 from math import sin, cos, pi
+from Button import *
 
 # Initialize Pygame
 pygame.init()
@@ -232,20 +233,7 @@ def home_page(screen, font):
 
 
 #training phase
-class Button:
-    def __init__(self, text, x, y):
-        self.text = text
-        self.x = x
-        self.y = y
-        self.font = pygame.font.Font(None, 36)
-        self.text_surf = self.font.render(self.text, True, (255, 255, 255))
-        self.rect = self.text_surf.get_rect(center=(self.x, self.y))
 
-    def draw(self, surface):
-        surface.blit(self.text_surf, self.rect)
-
-    def is_clicked(self, pos):
-        return self.rect.collidepoint(pos)
 
 # Initialize Start Game button
 start_button = Button("Start Game", screen_width / 2, screen_height / 2)
